pyproject.toml
[project]
name = "sirna-pre-process"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Saranya Pal", email = "saranya.pal@nuvo.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "viennarna>=2.7.0",
]

[project.scripts]
sirna-pre-process = "sirna_pre_process:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
